Re #INTOP-1470 A second version of WF26 upgrade

+++ opencell-security/src/main/java/org/meveo/security/UserAuthTimeProducer.java
@@ -34,7 +34,7 @@ public class UserAuthTimeProducer implements Serializable {
     /**
      * Timestamp when user has authenticated or token was issued
      */
-    protected int authenticatedAt;
+    protected Long authenticatedAt;
 
     /**
      * Authentication/session token hash/id
@@ -44,14 +44,14 @@ public class UserAuthTimeProducer implements Serializable {
     /**
      * @return Timestamp when user has authenticated or token was issued
      */
-    public int getAuthenticatedAt() {
+    public Long getAuthenticatedAt() {
         return authenticatedAt;
     }
 
     /**
      * @param authenticatedAt Timestamp when user has authenticated or token was issued
      */
-    public void setAuthenticatedAt(int authenticatedAt) {
+    public void setAuthenticatedAt(Long authenticatedAt) {
         this.authenticatedAt = authenticatedAt;
     }
